perf: keep the online numerator in BF16
Keep online max and sum in FP32, but store the recurrent attention numerator in BF16 because it feeds BF16 TensorE work and is ultimately returned as BF16. This halves the dominant numerator-state traffic.

Retest: 4 Trainium2 chips / 16 PCP ranks, LNC=2, global Q=1K, history=8K, FP8 KV, owner K transpose, 30 synchronized calls.
Host median: 4.176 ms -> 4.086 ms (-2.16%).
Device profile: 1.903 ms -> 1.821 ms (-4.30%).
Model MFU: 5.415% -> 5.658% (+0.243 pp), using 4 x 667 BF16 TFLOPS peak.

+++ pcp_attention/kernel.py
@@ -183,7 +183,10 @@ def make_history_attention_kernel(config: PCPAttentionConfig):
         )
         running_out = nl.zeros(
             (compute_rows, q_tiles, head_groups, d_tiles, 128),
-            dtype=nl.float32,
+            # The output numerator is consumed by BF16 TensorE matmuls and
+            # eventually returned as BF16. Keeping only max/sum in FP32 cuts
+            # the dominant recurrent vector-state traffic in half.
+            dtype=nl.bfloat16,
             buffer=nl.sbuf,
             name="online_numerator",
         )
@@ -473,18 +476,18 @@ def make_history_attention_kernel(config: PCPAttentionConfig):
                                 )
                                 pv_dq = nl.ndarray(
                                     (128, compute_rows),
-                                    dtype=nl.float32,
+                                    dtype=nl.bfloat16,
                                     buffer=nl.sbuf,
                                 )
                                 nisa.tensor_copy(dst=pv_dq, src=pv_psum)
                                 pv_qd = nl.ndarray(
                                     (compute_rows, 128),
-                                    dtype=nl.float32,
+                                    dtype=nl.bfloat16,
                                     buffer=nl.sbuf,
                                 )
                                 pv_qd_psum = nl.ndarray(
                                     (compute_rows, 128),
-                                    dtype=nl.float32,
+                                    dtype=nl.bfloat16,
                                     buffer=nl.psum,
                                 )
                                 nisa.nc_transpose(
